+++ test/endpoint.js
@@ -31,7 +31,7 @@ describe('phonetics', function() {
   require('./phonetics/caverphone.js');
   require('./phonetics/cologne.js');
   require('./phonetics/daitch-mokotoff.js');
-  require('./phonetics/doubleMetaphone.js');
+  require('./phonetics/double-metaphone.js');
   require('./phonetics/metaphone.js');
   require('./phonetics/mra.js');
   require('./phonetics/nysiis.js');